==> 01 data prep.ipynb <==
import warnings
warnings.filterwarnings("ignore")
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)
import lightgbm as lgb
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import confusion_matrix, roc_auc_score

### 1. importing dataset
* replace missing with ?

df = pd.read_excel('10k_diabetes.xlsx')
df = df.fillna('?')
df.shape

### 2. removing categorical columns
* multiple values (>20) - diag_3, diag_3_desc, diag_2, diag_2_desc, diag_1, diag_1_desc, discharge_disposition_id and medical_specialty
* single value (=1) - acetohexamide, citoglipton, examide, glimepiride.pioglitazone, metformin.pioglitazone, metformin.rosiglitazone and troglitazone
* highly skewed variable (tail count<30)- chlorpropamide, tolbutamide, miglitol, tolazamide and glipizide.metformin

df = df.drop(['diag_3','diag_3_desc','diag_2','diag_2_desc','diag_1','diag_1_desc',
              'discharge_disposition_id','medical_specialty'], axis=1)
df = df.drop(['acetohexamide','citoglipton','examide','glimepiride.pioglitazone',
              'metformin.pioglitazone','metformin.rosiglitazone','troglitazone'], axis=1)
df = df.drop(['chlorpropamide','tolbutamide','miglitol','tolazamide','glipizide.metformin'], axis=1)

df['age'] = np.where(df['age'].isin(['[10-20)','[0-10)']), 'Others', df['age'])
df['weight'] = np.where(df['weight'].isin(['[125-150)','[25-50)','[0-25)','[150-175)']), 'Others', df['weight'])
df['admission_type_id'] = np.where(df['admission_type_id'].isin(['Not Mapped','Newborn']), 'Others', df['admission_type_id'])
df['admission_source_id'] = np.where(df['admission_source_id'].isin(['HMO Referral','Not Mapped','Not Available','Court/Law Enforcement']), 'Others', df['admission_source_id'])
df['payer_code'] = np.where(df['payer_code'].isin(['WC','SI','CH','OT']), 'Others', df['payer_code'])
df['repaglinide'] = np.where(df['repaglinide']=='No', df['repaglinide'], 'Others')
df['nateglinide'] = np.where(df['nateglinide']=='No', df['nateglinide'], 'Others')
df['glimepiride'] = np.where(df['glimepiride'].isin(['Up','Down']), 'Others', df['glimepiride'])
df['pioglitazone'] = np.where(df['pioglitazone'].isin(['Up','Down']), 'Others', df['pioglitazone'])
df['rosiglitazone'] = np.where(df['rosiglitazone'].isin(['Up','Down']), 'Others', df['rosiglitazone'])
df['acarbose'] = np.where(df['acarbose']=='No', df['acarbose'], 'Others')
df['glyburide.metformin'] = np.where(df['glyburide.metformin']=='No', df['glyburide.metformin'], 'Others')
df.shape

### 3. removing numerical columns
* num_procedures has opposite trend (corr is -1)
* there are no missings in numerical variables
* all numerical variables have positive relationship with readmitted

df = df.drop(['num_procedures'], axis=1)
df['readmitted'] = np.where(df['readmitted']==False, 0, 1)
df[['readmitted','time_in_hospital','num_lab_procedures','num_medications',
    'number_outpatient','number_emergency','number_inpatient','number_diagnoses']].corr()

### 4. removing missing columns
* % missing > 50% 
* dropped weight, max_glu_serum, A1Cresult and payer_code

df = df.drop(['weight','max_glu_serum','A1Cresult','payer_code'], axis=1)
df.shape

for i in df.columns:
    print(i,' : ',df[df[i]=='?'].shape)

### 5. exporting dataset
* there are 7 numerical variables
* there are 18 categorical variables
* there are 1 target variable - readmitted
* also looking at univariate

df = df[['race','gender','age','admission_type_id','admission_source_id','metformin','repaglinide',
         'nateglinide','glimepiride','glipizide','glyburide','pioglitazone','rosiglitazone','acarbose',
         'insulin','glyburide.metformin','change','diabetesMed','time_in_hospital','num_lab_procedures',
         'num_medications','number_outpatient','number_emergency','number_inpatient','number_diagnoses',
         'readmitted']]
df.to_csv('01 data prep.csv', index=False)
df.shape

df.describe(include='all').T
